chore: implemented universal header for all pages to reduce component amount
