+++ src/app/menu/page.tsx
@@ -1,7 +1,7 @@
 // src/app/menu/page.tsx
 import { Montserrat } from "next/font/google";
-import MenuHeader from "#/src/components/menu/MenuHeader";
 import MenuManagement from "../../components/menu/MenuManagement";
+import UniversalHeader from "@/components/ui/UniversalHeader";
 
 const montserrat = Montserrat({
   subsets: ["latin"],
@@ -13,7 +13,7 @@ export default function MenuManagementPage() {
 
   return (
     <div className={`flex h-screen flex-col bg-[#F9F1E9] ${fontClassName}`}>
-      <MenuHeader pageName1="Menu" pageName2="Management" />
+      <UniversalHeader pageName1="Menu" pageName2="Management" />
       <MenuManagement fontClassName={fontClassName} />
     </div>
   );
